requirements.txt
# requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
langchain==0.0.340
langchain-text-splitters==0.0.1

# Vector embeddings and ML
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.24.4

# Database and Supabase
supabase==2.0.2
psycopg2-binary==2.9.9
pgvector==0.2.4

# LiveKit agent dependencies
livekit-agents==0.8.4
livekit-plugins-assemblyai==0.6.2
livekit-plugins-google==0.6.1
livekit-plugins-spitch==0.6.1
livekit-plugins-silero==0.6.1
livekit-plugins-noise-cancellation==0.8.0

# Additional utilities
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.1
pillow==10.1.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0